Implemented Meteostat to display historical temperature
Used meteostat to retrieve temperature data over the past 30 days and used plotly to visualise the data using a graph.

However there is a known bug where the current time can be incorrect... Although the implementation did not change it seems it may have not been fixed to begin with.